Added methods of listing entities in list entities module and import the moduel on app.rb
