+++ modules/list_entities.rb
@@ -1,2 +1,31 @@
+require_relative '../classroom'
+require_relative '../person'
+require_relative '../student'
+require_relative '../teacher'
+require_relative '../book'
+require_relative '../rental'
 module ListEntities
+  def list_books
+    @books.each { |book| puts "Title: #{book.title}, Author: #{book.author}" }
+  end
+
+  def list_people
+    @people.each { |person| puts "#{person.class}: Name: #{person.name}, Id: #{person.id}, Age: #{person.age}" }
+  end
+
+  def list_rentals_for_person
+    print 'Enter person id: '
+    person_id = gets.chomp.to_i
+
+    rentals = @rentals.select { |rental| rental.person.id == person_id }
+
+    if rentals.empty?
+      puts 'No rentals found for the given person id.'
+    else
+      puts 'Rentals:'
+      rentals.each do |rental|
+        puts "Date: #{rental.date}, Book: #{rental.book.title} by #{rental.book.author}"
+      end
+    end
+  end
 end
